remove printinfopass
