+++ test/arcilator/info-gathering-errors.mlir
@@ -1,4 +1,4 @@
-// RUN: circt-opt %s --arc-print-state-info --verify-diagnostics --split-input-file
+// RUN: arcilator %s --verify-diagnostics --split-input-file
 
 arc.model "Foo" {
 ^bb0(%arg0: !arc.storage<42>):
